+++ unity/MRMapper/Assets/Meshes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4a3ad4db3447e7047ab9dda589721f48
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
